requirements.txt
customtkinter
SpeechRecognition
pyaudio
openai
python-dotenv
psutil
edge-tts
pygame